Fix: Sofa tests
Former-commit-id: a9563222dd4cdce7362182882e793c72f2f368dc

+++ framework/sofa/core/objectmodel/Base.cpp
@@ -459,8 +459,8 @@ bool Base::parseField( const std::string& attribute, const std::string& value)
             ok = false;
         }
         sout << "Link " << linkVec[l]->getName() << " = " << linkVec[l]->getValueString() << sendl;
-        size_t s = linkVec[l]->getSize();
-        for (unsigned int i=0; i<(unsigned int)s; ++i)
+        unsigned int s = linkVec[l]->getSize();
+        for (unsigned int i=0; i<s; ++i)
         {
             sout  << "  " << linkVec[l]->getLinkedPath(i) << " = ";
             Base* b = linkVec[l]->getLinkedBase(i);
