+++ framework/sofa/core/objectmodel/BaseLink.h
@@ -127,10 +127,10 @@ public:
     /// This can be used to efficiently detect changes
     int getCounter(const core::ExecParams* params) const { return m_counters[core::ExecParams::currentAspect(params)]; }
 
-    virtual size_t getSize() const = 0;
-    virtual Base* getLinkedBase(size_t index=0) const = 0;
-    virtual BaseData* getLinkedData(size_t index=0) const = 0;
-    virtual std::string getLinkedPath(size_t index=0) const = 0;
+    virtual unsigned int getSize() const = 0;
+    virtual Base* getLinkedBase(unsigned int index=0) const = 0;
+    virtual BaseData* getLinkedData(unsigned int index=0) const = 0;
+    virtual std::string getLinkedPath(unsigned int index=0) const = 0;
 
     /// @name Serialization API
     /// @{
